+++ storm-core/src/jvm/org/apache/storm/utils/ConfigUtils.java
@@ -137,7 +137,7 @@ public class ConfigUtils {
 
     // public static mkStatsSampler // depends on Utils.evenSampler() TODO, this is sth we need to do after util
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static Map readStormConfig() {
         return _instance.readStormConfigImpl();
     }
@@ -243,7 +243,7 @@ public class ConfigUtils {
         return (masterLocalDir(conf) + FILE_SEPARATOR + "inimbus");
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static String supervisorLocalDir(Map conf) throws IOException {
         return _instance.supervisorLocalDirImpl(conf);
     }
@@ -258,7 +258,7 @@ public class ConfigUtils {
         return (supervisorLocalDir(conf) + FILE_SEPARATOR + "isupervisor");
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static String supervisorStormDistRoot(Map conf) throws IOException {
         return _instance.supervisorStormDistRootImpl(conf);
     }
@@ -267,7 +267,7 @@ public class ConfigUtils {
         return stormDistPath(supervisorLocalDir(conf));
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static String supervisorStormDistRoot(Map conf, String stormId) throws IOException {
         return _instance.supervisorStormDistRootImpl(conf, stormId);
     }
@@ -307,7 +307,7 @@ public class ConfigUtils {
         return (concatIfNotNull(stormRoot) + FILE_SEPARATOR + RESOURCES_SUBDIR);
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static LocalState supervisorState(Map conf) throws IOException {
         return _instance.supervisorStateImpl(conf);
     }
@@ -316,7 +316,7 @@ public class ConfigUtils {
         return new LocalState((supervisorLocalDir(conf) + FILE_SEPARATOR + "localstate"));
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static LocalState nimbusTopoHistoryState(Map conf) throws IOException {
         return _instance.nimbusTopoHistoryStateImpl(conf);
     }
@@ -325,7 +325,7 @@ public class ConfigUtils {
         return new LocalState((masterLocalDir(conf) + FILE_SEPARATOR + "history"));
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static Map readSupervisorStormConf(Map conf, String stormId) throws IOException {
         return _instance.readSupervisorStormConfImpl(conf, stormId);
     }
@@ -388,7 +388,7 @@ public class ConfigUtils {
         return ret;
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static void setWorkerUserWSE(Map conf, String workerId, String user) throws IOException {
         _instance.setWorkerUserWSEImpl(conf, workerId, user);
     }
@@ -409,7 +409,7 @@ public class ConfigUtils {
         new File(workerUserFile(conf, workerId)).delete();
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static String workerArtifactsRoot(Map conf) {
         return _instance.workerArtifactsRootImpl(conf);
     }
@@ -455,7 +455,7 @@ public class ConfigUtils {
         return new File((logRoot + FILE_SEPARATOR + id + FILE_SEPARATOR + port));
     }
 
-    // we use this "wired" wrapper pattern temporarily for mocking in clojure test
+    // we use this "weird" wrapper pattern temporarily for mocking in clojure test
     public static String workerRoot(Map conf) {
         return _instance.workerRootImpl(conf);
     }